feat: 导入自制谱 + 复制导出菜单 + 清理旧转谱代码
- 新增 ImportMusicModal：showDirectoryPicker 扫描文件夹导入自制谱
- 新增 ImportMusicCheck/ImportMusicExecute 后端 API（UGC/SUS→C2S + WAV→HCA→AWB + PNG→DDS）
- 新增 ExportOpt/ExportUgc/OpenExplorer/OpenXml 后端 API
- 新增 DropMenu「复制与导出」下拉菜单（导出 ZIP / 导出 UGC / 打开目录 / 编辑 XML）
- 删除旧转谱代码：C2sChart、UgcConverter、UgcParser、UmiguriService、UgcToolService（已被 MuConvert 子模块取代）
- 修复 i18n YAML 重复 key 导致前端编译失败
- 优化 FumenData 注释

+++ ChuChartManager/Models/FumenData.cs
@@ -24,10 +24,11 @@ public class FumenData
     public string NotesDesigner { get; set; } = "";
     public float DefaultBpm { get; set; }
 
-    // 实际等级 = level + levelDecimal/100，例如 level=12 decimal=70 → 12.7
+    // 谱面定数
     [JsonIgnore]
-    public double LevelValue => Level + LevelDecimal / 100.0;
+    public decimal LevelValue => Level + LevelDecimal / 100m;
 
+    // 等级显示
     public string LevelDisplay
     {
         get
